+++ site/components/element/tag.tsx
@@ -1,9 +1,9 @@
 import type { ComponentFn, Store } from "qrilljs/core";
 import { component, element } from "qrilljs/core";
-import { tag_map } from "../../site.config.ts";
+import { type TagSlugType, tag_map } from "../../site.config.ts";
 
 export type TagArgument = {
-    slug: string;
+    slug: TagSlugType;
 };
 
 export function tag(store: Store): ComponentFn<TagArgument> {