+++ site/pages/tags/[tag].html.tsx
@@ -1,9 +1,10 @@
 import type { RootPageFn, Store } from "qrilljs/core";
 import { element, registerRootPage, S_MEDIUM, style, W_MEDIUM } from "qrilljs/core";
 import { getAllMarkdowns } from "qrilljs/server";
+import * as v from "valibot";
 import { page } from "../../components/pages/page.tsx";
 import { summaries } from "../../components/sections/summaries.tsx";
-import { navitem, postFmSchema, posts_dir, site, tag_map } from "../../site.config.ts";
+import { navitem, postFmSchema, posts_dir, site, tag_map, tagSlugSchema } from "../../site.config.ts";
 
 type RootParameter = {
     tag: string;
@@ -30,7 +31,8 @@ export default function Root(store: Store): RootPageFn<RootParameter> {
     registerRootPage(store, styles);
 
     return async ({ tag }) => {
-        const md = (await getAllMarkdowns(posts_dir, postFmSchema)).filter((x) => x.data.tag?.includes(tag));
+        const parsed = v.parse(tagSlugSchema, tag);
+        const md = (await getAllMarkdowns(posts_dir, postFmSchema)).filter((x) => x.data.tag?.includes(parsed));
 
         return (
             <Page